renamed RDFTermFactory to RDFContext and introduced mechanism to get context from objects created by RDFContext factory.
This context sets the foundations for fluent syntax that lets us write something like

graph.addTriple(someSubject,integerDomainPredicate,12)

the context could also be subclassed if somebody wants to add functionality similar to the model-related objects.  If the extra pointers are wasting space,  one can set up the factory to be a singleton that is statically referenced.